acabar reportes movil

+++ movil/lib/classes/IMCGeneral.dart
@@ -1,5 +1,5 @@
 class IMCGeneral {
-  final int porcentaje;
+  final double porcentaje;
   final String label;
   IMCGeneral(this.porcentaje, this.label);
